+++ Cargo.toml
@@ -15,7 +15,8 @@ members = [
     "kitsune-http-signatures",
     "kitsune-messaging",
     "kitsune-search",
-    "kitsune-search/proto",
+    "kitsune-search-server",
+    "kitsune-search-server/proto",
     "kitsune-storage",
     "kitsune-type",
     "post-process",
