+++ kitsune-search-server/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "kitsune-search-server"
+version.workspace = true
+edition = "2021"
+
+[dependencies]
+autometrics = { version = "0.5.0", default-features = false, features = [
+    "metrics",
+] }
+dotenvy = "0.15.7"
+envy = "0.4.2"
+futures-util = "0.3.28"
+human-size = { version = "0.4.2", features = ["serde"] }
+kitsune-search-proto = { path = "proto" }
+metrics = "0.21.0"
+metrics-exporter-prometheus = "0.12.1"
+metrics-tracing-context = "0.14.0"
+metrics-util = "0.15.0"
+mimalloc = "0.1.37"
+prost = "0.11.9"
+serde = { version = "1.0.163", features = ["derive"] }
+tantivy = "0.19.2"
+time = "0.3.21"
+tokio = { version = "1.28.2", features = ["full"] }
+tonic = "0.9.2"
+tonic-health = "0.9.2"
+tower-http = { version = "0.4.0", features = ["add-extension", "trace"] }
+tracing = "0.1.37"
+tracing-subscriber = "0.3.17"
+
+[features]
+
+[dev-dependencies]
+rand = "0.8.5"
+tempfile = "3.5.0"
