+++ kitsune-search-server/src/grpc/service/index.rs
