+++ kitsune-search-server/src/grpc/service/mod.rs
