+++ kitsune-search-server/src/grpc/service/search.rs
