+++ kitsune-search-server/src/search/schema.rs
