+++ kitsune-search/Cargo.toml
@@ -1,35 +1,30 @@
 [package]
 name = "kitsune-search"
-version.workspace = true
 edition = "2021"
+version.workspace = true
 
 [dependencies]
-autometrics = { version = "0.5.0", default-features = false, features = [
-    "metrics",
-] }
-dotenvy = "0.15.7"
-envy = "0.4.2"
+async-trait = "0.1.68"
+diesel = "2.1.0"
+diesel-async = "0.3.0"
+diesel_full_text_search = { version = "2.1.0", default-features = false }
+enum_dispatch = "0.3.11"
 futures-util = "0.3.28"
-human-size = { version = "0.4.2", features = ["serde"] }
-kitsune-search-proto = { path = "proto" }
-metrics = "0.21.0"
-metrics-exporter-prometheus = "0.12.1"
-metrics-tracing-context = "0.14.0"
-metrics-util = "0.15.0"
-mimalloc = "0.1.37"
-prost = "0.11.9"
+kitsune-db = { path = "../kitsune-db" }
 serde = { version = "1.0.163", features = ["derive"] }
-tantivy = "0.19.2"
-time = "0.3.21"
-tokio = { version = "1.28.2", features = ["full"] }
-tonic = "0.9.2"
-tonic-health = "0.9.2"
-tower-http = { version = "0.4.0", features = ["add-extension", "trace"] }
+strum = { version = "0.24.1", features = ["derive"] }
+thiserror = "1.0.40"
+uuid = "1.3.3"
+
+# "kitsune-search" feature
+kitsune-search-proto = { path = "../kitsune-search-server/proto", optional = true }
+tonic = { version = "0.9.2", optional = true }
+
+# "meilisearch" feature
+meilisearch-sdk = { version = "0.23.2", optional = true }
 tracing = "0.1.37"
-tracing-subscriber = "0.3.17"
 
 [features]
-
-[dev-dependencies]
-rand = "0.8.5"
-tempfile = "3.5.0"
+default = []
+kitsune-search = ["dep:kitsune-search-proto", "dep:tonic"]
+meilisearch = ["dep:meilisearch-sdk"]
