+++ kitsune/Cargo.toml
@@ -47,6 +47,7 @@ kitsune-db = { path = "../kitsune-db" }
 kitsune-http-client = { path = "../kitsune-http-client" }
 kitsune-http-signatures = { path = "../kitsune-http-signatures" }
 kitsune-messaging = { path = "../kitsune-messaging" }
+kitsune-search = { path = "../kitsune-search" }
 kitsune-storage = { path = "../kitsune-storage" }
 kitsune-type = { path = "../kitsune-type" }
 mimalloc = "0.1.37"
@@ -103,13 +104,6 @@ async-graphql = { version = "5.0.9", default-features = false, features = [
 ], optional = true }
 async-graphql-axum = { version = "5.0.9", optional = true }
 
-# "kitsune-search" feature
-kitsune-search-proto = { path = "../kitsune-search/proto", optional = true }
-tonic = { version = "0.9.2", optional = true }
-
-# "meilisearch" feature
-meilisearch-sdk = { version = "0.23.2", optional = true }
-
 # "metrics" feature
 axum-prometheus = { version = "0.3.3", optional = true }
 metrics = { version = "0.21.0", optional = true }
@@ -127,9 +121,9 @@ serial_test = "2.0.0"
 [features]
 default = ["graphql-api", "kitsune-search", "mastodon-api", "metrics"]
 graphql-api = ["dep:async-graphql", "dep:async-graphql-axum"]
-kitsune-search = ["dep:kitsune-search-proto", "dep:tonic"]
+kitsune-search = ["kitsune-search/kitsune-search"]
 mastodon-api = []
-meilisearch = ["dep:meilisearch-sdk"]
+meilisearch = ["kitsune-search/meilisearch"]
 metrics = [
     "autometrics/metrics",
     "dep:axum-prometheus",
